Anim: fix null pointer access in 'Push-Down to NLA' operator
After pushing down the Action to the NLA, it needs to be tagged for
reevaluation. The code was using `adt->action->id` for this, but since
the nature of the operator is that it un-assigns the Action itself, that
would always be a null pointer.

Pull Request: https://projects.blender.org/blender/blender/pulls/128707